requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
jinja2==3.1.6
itsdangerous==2.2.0
certifi==2025.8.3
python-multipart==0.0.20
requests==2.32.5
beautifulsoup4==4.13.4
openai==1.107.2
pydantic==2.11.7
pypdf==6.0.0
python-docx==1.2.0
python-dotenv==1.1.1
